Added real Sleep command
Changed meaningless commands to better understandable commands
- old
  - haon
  - haoff
  - ppon
  - ppoff
- new
  - LetMePlay
  - TakeOver

Since the "ShutDown" and "ResetDay" commands work somewhat differently,
a "Sleep" command has been added.

The Sleep class has been modified. The commands are now in the base class,
which functions as a worker, but can still be called via the derived class.